% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledgeCategory.R
\name{kc_create}
\alias{kc_create}
\title{创建知识库}
\usage{
kc_create(parentName, name)
}
\arguments{
\item{parentName}{上级分类名称}

\item{name}{本级名称}
}
\value{
返回值
}
\description{
创建知识库
}
\examples{
kc_create()


}
