% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledgeCategory.R
\name{kc_upload}
\alias{kc_upload}
\title{将知识库上传到数据库中}
\usage{
kc_upload(conn = conn_rds("nsic"), data = kc_query(), table_name = "t_ksm_kc")
}
\arguments{
\item{conn}{数据库连接}

\item{data}{知识分类数据}

\item{table_name}{数据库表名称}
}
\value{
无
}
\description{
将知识库上传到数据库中
}
\examples{
 kc_upload()
 create table t_ksm_kc
(FName varchar(200),
 FId varchar(50),
 FParentId varchar(50),
 primary key(FId))
}
