% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledgeNode.R
\name{kn_create}
\alias{kn_create}
\title{创建知识点}
\usage{
kn_create(txt, kc_name, conn = conn_rds("nsic"))
}
\arguments{
\item{txt}{知识点名称}

\item{kc_name}{知识点分类}

\item{conn}{数据库连接}
}
\value{
返回值
}
\description{
创建知识点
}
\examples{
kc_create()
kn_create('we can do it','l112')
如果内容已经存在，则系统会报错，
目前在python中没有做处理，需要在R中进行处理
在下一个版本中进行改进，返回所有的内容，并更新知识库
}
