% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledgeNode.R
\name{kn_is_exist}
\alias{kn_is_exist}
\title{判断知识点是否存在}
\usage{
kn_is_exist(kn_name, kc_name, conn = conn_rds("nsic"))
}
\arguments{
\item{kn_name}{知识点名称}

\item{kc_name}{知识分类名称}

\item{conn}{连接}
}
\value{
返回值
}
\description{
判断知识点是否存在
}
\examples{
kn_is_exist()
}
