% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledgeNode.R
\name{kn_upload}
\alias{kn_upload}
\title{知识点上传}
\usage{
kn_upload(
  conn = conn_rds("nsic"),
  data = kn_queryAll(kc_name = "l112"),
  table_name = "t_ksm_kn"
)
}
\arguments{
\item{conn}{数据库连接}

\item{data}{知识分类的数据}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
知识点上传
}
\examples{
kn_upload()
}
